Expose manual split in verification task
